Added a login checker and moved the log printstream
Added a login checker to make sure the user has put in the proper login info (fixes #15, ref #19). Also moved the logging print stream to main to allow for more distribution for loggging if we want it. Removed empty pin and uid from logging as it should never make it there.