Adds a semester/date test for schedule now button
This push adds a test for semester and date for the schedule now button. (ref #19.) If you put in a semester that is not yet able to be scheduled it will throw an error because the semester doesn’t exist yet in the semester selection. If you throw in a previous semester that does not have CRN boxes anymore (i.e. past register date) it will also throw an error.