Removes the old semester check
The old semester check is unessecary and will not work for scheduling later. Changed to a try catch so we will not be checking for an error in your semester selection. closes #19. I will add a date check to make sure you don't do a date that has always passed soon.